refactor: moved lib/ to utils/
